Convert LumiNUS auth provider to ADFS auth provider
