+++ config/dev.secrets.exs.example
@@ -3,19 +3,11 @@ import Config
 config :cadet,
   identity_providers: %{
     # If you are just doing development, you can simply use the test provider
-    # below. You only need to use LumiNUS or OpenID if you are deploying to
+    # below. You only need to use ADFS or OpenID if you are deploying to
     # production.
 
-    # # To use authentication with LumiNUS.
-    # "nusnet_id" =>
-    #   {Cadet.Auth.Providers.LumiNUS,
-    #    %{
-    #      # The LumiNUS API key
-    #      api_key: "",
-    #      # The modules to check for enrolment. Multiple modules can be specified,
-    #      # in which case the highest role from any of the modules will be used.
-    #      modules: %{"CS1101S" => "2010"}
-    #    }},
+    # # To use authentication with NUS's ADFS instance.
+    # "nusnet_id" => {Cadet.Auth.Providers.ADFS, %{token_endpoint: "https://my-adfs/adfs/oauth2/token"}},
     # # An example of OpenID authentication with Cognito. Any OpenID-compliant
     # # provider should work.
     # "cognito" =>
