+++ test/fixtures/custom_cassettes/adfs/authorise#1.json
@@ -1,0 +1,28 @@
+[
+  {
+    "request": {
+      "body": "client_id=CLIENT_ID&code=CODE&grant_type=authorization_code&redirect_uri=http%3A%2F%2Flocalhost%3A8000%2Flogin",
+      "headers": {
+        "Content-Type": "application/x-www-form-urlencoded"
+      },
+      "method": "post",
+      "options": [],
+      "request_body": "",
+      "url": "https://my-adfs/adfs/oauth2/token"
+    },
+    "response": {
+      "binary": false,
+      "body": "{\"access_token\":\"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJTYW1BY2NvdW50TmFtZSI6ImUwMTIzNDU2IiwiZGlzcGxheU5hbWUiOiJZb3VyIE5hbWUiLCJpYXQiOjE1MTYyMzkwMjIsImV4cCI6MjAwMDAwMDAwMH0.YkXG5snv_aMQtctrhrdHIrJbt0gmDABuUV-7YJ4FCHM2G_b6N_cEWssUsaiTEsjJzTCVvTdMdgH4q4jqfAZ75-h9WTDv6erDk7Uee8k0HJ4Gi08gX7u_efRcqLRi2ydv1ed74LCoox_SLi97C5tYZBTJwMI6Ljm1HIO4VVGwbZpDXTjxwqvMzUw0bxDYtPgVhU-PE79rcvmGNuaWzt5GloQl6hgVYWtJpCbKh_fTT_d5czsq0TWXsCwSY9OK96ho966PrryjBSAvfjSFD4rNUb2c8vDKO1ozjMEpDBWgvjdUJXi4rdTdqFny-bgFG2dFZksIgnAZxaryH3m0AcIUOg\",\"token_type\":\"bearer\",\"expires_in\":3600}",
+      "headers": {
+        "Cache-Control": "no-store",
+        "Pragma": "no-cache",
+        "Content-Length": "558",
+        "Content-Type": "application/json;charset=UTF-8",
+        "Server": "Microsoft-HTTPAPI/2.0 Microsoft-HTTPAPI/2.0",
+        "Date": "Sun, 03 Jul 2022 07:13:56 GMT"
+      },
+      "status_code": 200,
+      "type": "ok"
+    }
+  }
+]
