+++ test/fixtures/custom_cassettes/adfs/authorise#5.json
@@ -1,0 +1,28 @@
+[
+  {
+    "request": {
+      "body": "blah",
+      "headers": {
+        "Content-Type": "application/x-www-form-urlencoded"
+      },
+      "method": "post",
+      "options": [],
+      "request_body": "",
+      "url": "https://my-adfs/adfs/oauth2/token"
+    },
+    "response": {
+      "binary": false,
+      "body": "{\"error\":\"invalid_client\",\"error_description\":\"boom.\"}",
+      "headers": {
+        "Cache-Control": "no-store",
+        "Pragma": "no-cache",
+        "Content-Length": "173",
+        "Content-Type": "application/json;charset=UTF-8",
+        "Server": "Microsoft-HTTPAPI/2.0 Microsoft-HTTPAPI/2.0",
+        "Date": "Sun, 03 Jul 2022 07:17:08 GMT"
+      },
+      "status_code": 500,
+      "type": "ok"
+    }
+  }
+]
