--- test/fixtures/custom_cassettes/luminus/authorise#1.json
@@ -1,27 +1,0 @@
-[
-  {
-    "request": {
-      "body": "client_id=CLIENT_ID&code=CODE&grant_type=authorization_code&redirect_uri=REDIRECT_URI",
-      "headers": {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Ocp-Apim-Subscription-Key": "API_KEY"
-      },
-      "method": "post",
-      "options": [],
-      "request_body": "",
-      "url": "https://luminus.azure-api.net/login/adfstoken"
-    },
-    "response": {
-      "binary": false,
-      "body": "{\"access_token\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzYW1BY2NvdW50TmFtZSI6InVzZXJuYW1lIn0.ovrGQZr-X4IS4tHgbwc2ESm-cevOXnNVU5BcK3Z1bsA\",\"token_type\":\"bearer\",\"expires_in\":86400}",
-      "headers": {
-        "Content-Length": "180",
-        "Content-Type": "application/json; charset=utf-8",
-        "X-Powered-By": "ASP.NET",
-        "Date": "Wed, 27 May 2020 12:25:48 GMT"
-      },
-      "status_code": 200,
-      "type": "ok"
-    }
-  }
-]
