--- test/fixtures/custom_cassettes/luminus/authorise#2.json
@@ -1,27 +1,0 @@
-[
-  {
-    "request": {
-      "body": "client_id=CLIENT_ID&code=CODE_invalid&grant_type=authorization_code&redirect_uri=REDIRECT_URI",
-      "headers": {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Ocp-Apim-Subscription-Key": "API_KEY"
-      },
-      "method": "post",
-      "options": [],
-      "request_body": "",
-      "url": "https://luminus.azure-api.net/login/adfstoken"
-    },
-    "response": {
-      "binary": false,
-      "body": "{\"error\":\"invalid_grant\",\"error_description\":\"MSIS9612: The authorization code received in \\u0027code\\u0027 parameter is invalid. \"}",
-      "headers": {
-        "Transfer-Encoding": "chunked",
-        "Content-Type": "text/plain; charset=utf-8",
-        "X-Powered-By": "ASP.NET",
-        "Date": "Wed, 27 May 2020 12:28:38 GMT"
-      },
-      "status_code": 200,
-      "type": "ok"
-    }
-  }
-]
