--- test/fixtures/custom_cassettes/luminus/authorise#3.json
@@ -1,27 +1,0 @@
-[
-  {
-    "request": {
-      "body": "client_id=CLIENT_ID&code=CODE&grant_type=authorization_code&redirect_uri=REDIRECT_URI_invalid",
-      "headers": {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Ocp-Apim-Subscription-Key": "API_KEY"
-      },
-      "method": "post",
-      "options": [],
-      "request_body": "",
-      "url": "https://luminus.azure-api.net/login/adfstoken"
-    },
-    "response": {
-      "binary": false,
-      "body": "{\"error\":\"invalid_request\",\"error_description\":\"MSIS9609: The \\u0027redirect_uri\\u0027 parameter is invalid. No redirect uri with the specified value is registered for the received \\u0027client_id\\u0027. \"}",
-      "headers": {
-        "Transfer-Encoding": "chunked",
-        "Content-Type": "text/plain; charset=utf-8",
-        "X-Powered-By": "ASP.NET",
-        "Date": "Wed, 27 May 2020 12:27:52 GMT"
-      },
-      "status_code": 200,
-      "type": "ok"
-    }
-  }
-]
