--- test/fixtures/custom_cassettes/luminus/get_name#1.json
@@ -1,30 +1,0 @@
-[
-  {
-    "request": {
-      "body": "",
-      "headers": {
-        "Ocp-Apim-Subscription-Key": "API_KEY",
-        "Authorization": "Bearer TOKEN"
-      },
-      "method": "get",
-      "options": {
-        "recv_timeout": 10000,
-        "connect_timeout": 10000
-      },
-      "request_body": "",
-      "url": "https://luminus.azure-api.net/user/Profile"
-    },
-    "response": {
-      "binary": false,
-      "body": "{\"userNameOriginal\":\"TEST LUMINUS USER\"}",
-      "headers": {
-        "Content-Length": "40",
-        "Content-Type": "application/json; charset=utf-8",
-        "X-Powered-By": "ASP.NET",
-        "Date": "Wed, 27 May 2020 12:44:27 GMT"
-      },
-      "status_code": 200,
-      "type": "ok"
-    }
-  }
-]
