--- test/fixtures/custom_cassettes/luminus/get_role#4.json
@@ -1,32 +1,0 @@
-[
-  {
-    "request": {
-      "body": "",
-      "headers": {
-        "Ocp-Apim-Subscription-Key": "API_KEY",
-        "Authorization": "Bearer TOKEN"
-      },
-      "method": "get",
-      "options": {
-        "recv_timeout": 10000,
-        "connect_timeout": 10000
-      },
-      "request_body": "",
-      "url": "https://luminus.azure-api.net/module"
-    },
-    "response": {
-      "binary": false,
-      "body": "{\"status\":\"success\",\"code\":200,\"total\":6,\"offset\":0,\"data\":[{\"id\":\"41d42e9a-5880-43b5-8ee6-75f5a41355e3\",\"createdDate\":\"2020-04-26T09:14:31.927+08:00\",\"creatorID\":\"4dce39fc-432c-4e4c-aef1-8bb187fcbcc0\",\"lastUpdatedDate\":\"2020-04-29T14:07:19.017+08:00\",\"lastUpdatedBy\":\"878407b2-d201-40ca-a097-417abca07545\",\"name\":\"CS1101S\",\"startDate\":\"2000-01-01T00:00:00.000+08:00\",\"endDate\":\"2001-01-01T00:00:00.000+08:00\",\"publish\":false,\"parentID\":\"41d42e9a-5880-43b5-8ee6-75f5a41355e3\",\"resourceID\":\"00000000-0000-0000-0000-000000000000\",\"access\":{\"access_Full\":false,\"access_Read\":true,\"access_Create\":true,\"access_Update\":true,\"access_Delete\":true,\"access_Settings_Read\":true,\"access_Settings_Update\":true},\"courseName\":\"Programming Methodology\",\"facultyCode\":\"39\",\"departmentCode\":\"252\",\"term\":\"2010\",\"acadCareer\":\"UGRD\",\"courseSearchable\":true,\"allowAnonFeedback\":\"N\",\"displayLibGuide\":false,\"copyFromID\":\"57290e55-335a-4c09-b904-a795572d6cda\",\"l3\":false,\"enableLearningFlow\":true,\"usedNusCalendar\":false,\"isCorporateCourse\":false}]}",
-      "headers": {
-        "Cache-Control": "no-store,no-cache",
-        "Pragma": "no-cache",
-        "Content-Length": "1027",
-        "Content-Type": "application/json; charset=utf-8",
-        "X-Powered-By": "ASP.NET",
-        "Date": "Wed, 27 May 2020 12:44:21 GMT"
-      },
-      "status_code": 200,
-      "type": "ok"
-    }
-  }
-]
